+++ AffiseAttributionLib/Classes/parameters/providers/DensityProvider.swift
@@ -1,0 +1,23 @@
+/**
+ * Provider for parameter [ProviderType.DENSITY]
+ */
+class DensityProvider: FloatPropertyProvider {
+
+    private let useCase: ScreenUseCase
+
+    init(useCase: ScreenUseCase) {
+        self.useCase = useCase
+    }
+
+    override func provide() -> Double? {
+        return useCase.getDensity()
+    }
+
+    public override func getOrder() -> Float {
+        22.4
+    }
+
+    public override func getKey() -> ProviderType? {
+        ProviderType.DENSITY
+    }
+}