+++ AffiseAttributionLib/Classes/usecase/DiskUseCase.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+protocol DiskUseCase {
+    func getTotalDisk() -> Int64?
+
+    func getFreeDisk() -> Int64?
+}