+++ AffiseAttributionLib/Classes/usecase/ProcessInfoUseCase.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+protocol ProcessInfoUseCase {
+    func getCpuName() -> String
+    func getCpuCores() -> Int64
+}