+++ AffiseAttributionLib/Classes/usecase/ScreenUseCase.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+protocol ScreenUseCase {
+    func getScreenWidth() -> Int64
+
+    func getScreenHeight() -> Int64
+
+    func getDensity() -> Double
+}